Fixing a bug where multiple firewalls with authenticators would not work
It turns out that the AuthenticationProviderManager receives 2 instances of GuardAuthenticationProvider,
one for reach of the 2 firewalls. Each contains the guard authenticator(s) for each firewall.
So, if no matching authenticator is found, it's likely because the next GuardAuthenticationProvider
actually contains the correct guard authenticator. We need to not throw an error so the
next one can be tried.